Moved Client functions and header into their own files
